+++ src/main/java/com/trepicoder/adventurecraft/blocks/ACCampfire.java
@@ -1,0 +1,65 @@
+package com.trepicoder.adventurecraft.blocks;
+
+import com.trepicoder.adventurecraft.utils.VoxelShapeHelper;
+import net.minecraft.block.Block;
+import net.minecraft.block.BlockState;
+import net.minecraft.block.Blocks;
+import net.minecraft.block.HorizontalBlock;
+import net.minecraft.item.BlockItemUseContext;
+import net.minecraft.state.StateContainer;
+import net.minecraft.state.properties.BlockStateProperties;
+import net.minecraft.util.math.BlockPos;
+import net.minecraft.util.math.shapes.ISelectionContext;
+import net.minecraft.util.math.shapes.VoxelShape;
+import net.minecraft.util.math.shapes.VoxelShapes;
+import net.minecraft.world.IBlockReader;
+import net.minecraft.world.World;
+
+import javax.annotation.Nullable;
+import java.util.Random;
+
+public class ACCampfire extends HorizontalBlock {
+    public ACCampfire(Properties builder) {
+        super(builder);
+    }
+
+    @Override
+    public VoxelShape getShape(BlockState state, IBlockReader worldIn, BlockPos pos, ISelectionContext context) {
+        VoxelShape body1 = VoxelShapeHelper.createRotatedShape(state.get(BlockStateProperties.HORIZONTAL_FACING),0, 0, 5, 16,2,11);
+        VoxelShape body2 = VoxelShapeHelper.createRotatedShape(state.get(BlockStateProperties.HORIZONTAL_FACING), 5, 0, 0,11, 2, 16);
+
+        VoxelShape body3 = VoxelShapeHelper.createRotatedShape(state.get(BlockStateProperties.HORIZONTAL_FACING), 1, 0, 4, 15, 2, 12);
+        VoxelShape body4 = VoxelShapeHelper.createRotatedShape(state.get(BlockStateProperties.HORIZONTAL_FACING), 4, 0, 1, 12, 2, 15);
+
+        VoxelShape body5 = VoxelShapeHelper.createRotatedShape(state.get(BlockStateProperties.HORIZONTAL_FACING), 2, 0, 3, 14, 2, 13);
+        VoxelShape body6 = VoxelShapeHelper.createRotatedShape(state.get(BlockStateProperties.HORIZONTAL_FACING), 3, 0, 2, 13, 2, 14);
+
+        return VoxelShapes.or(body1, body2, body3, body4, body5, body6);
+    }
+
+    @Override
+    public boolean isSolid(BlockState state) {
+        return false;
+    }
+
+    @Override
+    public void animateTick(BlockState stateIn, World worldIn, BlockPos pos, Random rand) {
+        // Animate fire and smoke particles
+    }
+
+    @Nullable
+    @Override
+    public BlockState getStateForPlacement(BlockItemUseContext context) {
+        BlockState blockstate = super.getStateForPlacement(context);
+
+        if (blockstate != null) {
+            blockstate = blockstate.with(BlockStateProperties.HORIZONTAL_FACING, context.getPlacementHorizontalFacing().getOpposite());
+        }
+        return blockstate;
+    }
+
+    @Override
+    protected void fillStateContainer(StateContainer.Builder<Block, BlockState> container) {
+        container.add(BlockStateProperties.HORIZONTAL_FACING);
+    }
+}