+++ src/main/java/com/trepicoder/adventurecraft/blocks/ACChoppingBlock.java
@@ -1,0 +1,24 @@
+package com.trepicoder.adventurecraft.blocks;
+
+import net.minecraft.block.Block;
+import net.minecraft.block.BlockState;
+import net.minecraft.block.SoundType;
+import net.minecraft.block.material.Material;
+import net.minecraft.util.math.BlockPos;
+import net.minecraft.util.math.shapes.ISelectionContext;
+import net.minecraft.util.math.shapes.VoxelShape;
+import net.minecraft.util.math.shapes.VoxelShapes;
+import net.minecraft.world.IBlockReader;
+import net.minecraftforge.common.ToolType;
+
+public class ACChoppingBlock extends Block {
+    public ACChoppingBlock() {
+        super(Block.Properties.create(Material.WOOD).hardnessAndResistance(2.0f).sound(SoundType.WOOD).harvestTool(ToolType.AXE));
+    }
+
+    @Override
+    @SuppressWarnings("deprecation")
+    public VoxelShape getShape(BlockState state, IBlockReader worldIn, BlockPos pos, ISelectionContext context) {
+        return VoxelShapes.create(0, 0, 0, 16, 5, 16);
+    }
+}