+++ src/main/java/com/trepicoder/adventurecraft/registries/ACBlocks.java
@@ -1,15 +1,23 @@
 package com.trepicoder.adventurecraft.registries;
 
 import net.minecraft.block.Block;
+import net.minecraft.item.Item;
 
 public class ACBlocks {
     //region Wood
-    //region Logs
     public static Block sugarmaple_log;
+    public static Block sugarmaple_planks;
+    public static Block sugarmaple_leaves;
     //endregion
 
-    //region Planks
-    public static Block sugarmaple_planks;
-    //endregion
+    //region Tile Entities
+    public static Block campfire;
+    public static Block acacia_chopping_block;
+    public static Block birch_chopping_block;
+    public static Block dark_oak_chopping_block;
+    public static Block jungle_chopping_block;
+    public static Block oak_chopping_block;
+    public static Block spruce_chopping_block;
+    public static Block sugarmaple_chopping_block;
     //endregion
 }