+++ src/main/java/com/trepicoder/adventurecraft/registries/ACItems.java
@@ -1,5 +1,6 @@
 package com.trepicoder.adventurecraft.registries;
 
+import net.minecraft.client.renderer.color.BlockColors;
 import net.minecraft.item.Item;
 
 public class ACItems {
@@ -12,18 +13,31 @@ public class ACItems {
 
     //endregion
 
-    //===================================================================
-
     //region Misc
     public static Item plantfibres;
     public static Item twine;
     public static Item treesap;
+    public static Item stickpile;
     //endregion
+
+    //===================================================================
 
     //region Item Blocks
     //region Wood
     public static Item sugarmaple_log;
     public static Item sugarmaple_planks;
+    public static Item sugarmaple_leaves;
+    //endregion
+
+    //region Tile Entities
+    public static Item campfire;
+    public static Item acacia_chopping_block;
+    public static Item birch_chopping_block;
+    public static Item dark_oak_chopping_block;
+    public static Item jungle_chopping_block;
+    public static Item oak_chopping_block;
+    public static Item spruce_chopping_block;
+    public static Item sugarmaple_chopping_block;
     //endregion
     //endregion
 }