+++ changelog.md
@@ -1,5 +1,75 @@
 # Changelog
 
+## ALPHA 1.0.1
+
+### Blocks
+
+- Added Acacia Chopping Block
+- Added Birch Chopping Block
+- Added Dark Oak Chopping Block
+- Added Jungle Chopping Block
+- Added Oak Chopping Block
+- Added Spruce Chopping Block
+- Added Sugarmaple Chopping Block
+- Added Sugarmaple Leaves
+- Added Campfire
+
+### Items
+
+- Added Acacia Chopping Block
+- Added Birch Chopping Block
+- Added Dark Oak Chopping Block
+- Added Jungle Chopping Block
+- Added Oak Chopping Block
+- Added Spruce Chopping Block
+- Added Sugarmaple Chopping Block
+- Added Sugarmaple Leaves
+- Added Campfire
+- Added Stick Bundle
+
+### Loot Drops
+
+- Added loot drop for Acacia Chopping Block
+- Added loot drop for Birch Chopping Block
+- Added loot drop for Dark Oak Chopping Block
+- Added loot drop for Jungle Chopping Block
+- Added loot drop for Oak Chopping Block
+- Added loot drop for Spruce Chopping Block
+- Added loot drop for Sugarmaple Chopping Block
+- Added loot drop for Sugarmaple Leaves
+- Added loot drop for Campfire
+
+### Recipes
+
+- Added recipe for Acacia Chopping Block
+- Added recipe for Birch Chopping Block
+- Added recipe for Dark Oak Chopping Block
+- Added recipe for Jungle Chopping Block
+- Added recipe for Oak Chopping Block
+- Added recipe for Spruce Chopping Block
+- Added recipe for Sugarmaple Chopping Block
+- Added recipe for Campfire
+- Added recipe for Stick Bundle
+
+### GUI
+
+- Added gui texture for Chopping Block
+- Added gui texture for Wood Pile
+
+### WIP
+
+- Add tile entity for Acacia Chopping Block
+- Add tile entity for Birch Chopping Block
+- Add tile entity for Dark Oak Chopping Block
+- Add tile entity for Jungle Chopping Block
+- Add tile entity for Oak Chopping Block
+- Add tile entity for Spruce Chopping Block
+- Add tile entity for Sugarmaple Chopping Block
+- Add tile entity for Campfire
+- Add worldgen for Sugarmaple Tree
+
+---
+
 ## ALPHA 1.0.0
 
 ### Blocks